ChunkedSpaceMesh: use the same code for initial blocks as changed blocks.
This is a significant simplification, and will hopefully allow for clean
addition of the feature of incremental initialization rather than doing
all the work on the first frame.

Note that `todo.all_blocks_and_chunks` is now true on the first frame,
and must be or the mesh vector length adjustment won't be performed and
the per-chunk logic will panic.