Fix WebGL compatibility of info text shader.
The problem was passing an integer instead of float for `texture()`'s
`bias` parameter, but we didn't need that at all. Also added some
variables to make the code more descriptive.

Note: We don't actually need the info text shader at all, because
the web build uses HTML for the info text. But there's no code to
skip compiling it.