Add `GraphicsOptions::tone_mapping`.
This does not implement a *good* tone mapping operator, well integrated
with knowledge of the brightness and dynamic range of the scene. It's
intended to be a prototype setting up most of the necessary plumbing for
doing so later, probably in combination with probing the lighting around
the character.

Currently, callers of the raytracer have to to explicitly apply tone
mapping, which is unfortunate; I plan to improve that, perhaps by adding
an analogue to `trace_scene_to_text` which produces an explicitly RGB
image.