New volume allocator algorithm `Alloctree`.
This is going to replace the current texture allocator algorithm, but it
is currently only tested.

Its advantage is that it supports “tiles” of multiple sizes, rather than
a specific size set at creation time (fixed at 16³ texels in practice).
Currently, all allocations are cubes, but I intend to add support for
“flat” regions to make use cases like text drawing more efficient.