Use `cgmath::Decomposed` to define camera transforms.
This reduces the number of invalid cases and will allow converting
the transform into a `Spawn` configuration so we can share look-at code.
But for now, this commit only introduces the additional method and
changes the `Character` and `Vui` outputs.

Note that the new transform is inverted from the matrix definition:
it transforms eye coordinates to world coordinates, not vice versa.
This means that the transform's displacement is directly in world
coordinates.