Add `RotationPlacementRule::Attach` and `ToolInput::place_block`.
I expect `ToolInput::place_block` to become the entry point for all
special placement behaviors that might come in the future, like
multiblock structures and reacting to the neighborhood.

`Attach` in its current form is just the simplest useful placement rule
I could think of. Future work: specify the currently-unconstrained axis,
and specify what happens in the unsatisfiable cases.

Also derive `Arbitrary` for `Face` so it can be used in placement rules.