Split block data calc of `PixelBuf` into new trait `RtBlockData`.
Also, `RtBlockData`'s factory functions accept the graphics options and
a new custom options type, so block data can be customized for the
current options. This provision is not yet used.

The block data is no longer required to be `'static`; theoretically this
could reduce copying in `CharacterRtData` but I haven't tried that.