RaytraceToTexture: Disable pixel randomization.
This feels a lot better — not sure how much is reduction in noise and
how much (if any) is better performance due to more coherent memory
accesses.

(Yes, RaytraceToTexture is still "dead_code". I have a branch where
I'm testing it out as a hardcoded replacement, which needs to be made
into a regular graphics option.)