Add `-g print` mode which prints the scene and exits.
Unlike `record` (image writing) mode, it doesn't automatically enable
`--precompute-light`. This is because the idea was that this could be
used for “rapid prototyping” where the scene is displayed ASAP after
a code change. That's a thin rationale for user-facing behavior choice,
though, so maybe I should change it and also add a
`--no-precompute-light` option.