all-is-cubes-desktop: Add option to load a file.
I haven't yet implemented a “native” format, so for now it only
supports loading MagicaVoxel .vox format -- chosen because there is a
convenient library (or several, actually) and it seems to be a fairly
popular format for sharing voxel models.

Future work: The .vox importer should probably not live exclusively in
all-is-cubes-desktop, but I don't think it should go in the core library
crate and I'm not sure what to name a new crate for the purpose.