Move resource force-static kludge to all-is-cubes-wasm.
It turns out that Cargo unifies features across a workspace and
across targets, and the target-specific dependency in all-is-cubes
doesn't affect that. Putting the feature enabling in the wasm package
does, *if* we also run `cargo run --package all-is-cubes-desktop`
instead of `cargo run --bin all-is-cubes`. Awkward, but the worst
failure is that hot reloading doesn't work.

Future work: contribute a patch to `resource` so that it automatically
works on wasm.