Constify conversion from sRGB using lookup tables.
* `Rgba::from_srgb_32bit` is now `const fn`.
* Added `Rgb::from_srgb_32bit`.

This will allow defining constant colors in non-linear sRGB form, which
is the same convention used by CSS and nearly all other casual color
notation, so it will be possible to statically provide previews of those
same colors, and the constants themselves will be more familiar to
readers used to reading “hex code” colors.

A disadvantage is that these 8-bit-per-channel colors are no good for
HDR (so, for example, `Rgb::from_srgb_32bit` can't represent many
values that would be useful for lighting).

This commit does not include any helpers for concise notation like the
existing `rgba_const!()`.