Split GPU/luminance code into its own crate.
The primary goal of this is to improve build performance. It's only a
1-2 second savings (out of 6-7) on my current machine, but that's time
cut off of the analysis of `all-is-cubes`, which is the step that all
others depend on.

Secondarily, it's a well-defined unit of code that works out neatly as a
separate crate -- well, except for all the functions it uses that I was
keeping private, which I've now mostly changed to be public and
`#[doc(hidden)]`.

(In principle, it's a good principle to ensure the renderer has no
special access. In practice, there's lots of things that I'd rather not
declare “stable” yet, even more so than the rest of the project isn't
stable.)

`xtask` no longer has the `--without-luminance` option, but I do not
currently have a need for it and I'll revisit that if the issue comes
up again. (For example, it might make sense not to avoid `luminance`
but to avoid `glfw`, i.e. add a feature to `all-is-cubes-desktop` to
disable windowed mode).

The reason I called the crate `all-is-cubes-gpu` is because I expect
there might be reason to switch to a different graphics library in the
future, and I'd like to reuse the same crate structure (and *possibly*
parts like vertex types and sequence of rendering steps) rather than
defining a new crate and populating crates.io with defunct crate names.