Create `all_is_cubes_desktop::glue` for simple API glue functions.
This refactoring is aimed at eventually being able to better separate
the choice of window/events/platform API from the choice of what
operations are being performed; this makes there be a place for the
pure glue code to live, and be seen in parallel. (for example, if
`all_is_cubes::apps::Key` is expanded, then each of the glue modules
should be reviewed and updated).