Replace `strum::IntoEnumIterator` with `exhaust::Exhaust`.
This will allow `BlockModule`s to be defined using enums with fields.

Future work: we might choose to replace the `strum::Display` derive with
something else (and will need to, to take advantage of fields). However,
that doesn't change the public API of the crate, since the trait it
derives is just `std::fmt::Display`.