Add `#[must_use]` to various types and functions.
This was warned about by a new Clippy lint in Rust 1.59.0 (prerelease).
That lint is going to be moved to `clippy::pedantic`, but I think I want
to keep it anyway, so this addresses its occurrences and also explicitly
enables it.

It was not necessary to add `#[must_use]` to all `Transaction` types,
but I did it anyway.