Make wasm YieldProgress callbacks Send+Sync.
This will enable making YieldProgress itself Send+Sync.

For the yield callback, I replaced the use of gloo-timers with a
from-scratch implementation using a oneshot future. This also will be
reusable in the case I get around to replacing setTimeout() with a
better implementation (setTimeout() infamously has a built-in minimum
delay).

For the progress callback, we need to actually interact with the
!Send DOM element proxy, so I used SendWrapper.