Relocate fuzz package to workspace root.
Now that we have multiple library crates (-gpu and -content), it makes
sense to be able to add fuzzers for those. This also saves a `cd` to
run `cargo fuzz`.

There have been some issues with cargo-fuzz in workspaces, but the only
open one is with initializing, which we don't need to do again, and I
don't seem to have any trouble with running the fuzzer in this
configuration.