Make `Block` a pointer to "primitive" and "modifiers".
A "primitive" is one of the three base cases previously existent:
`Atom`, `Recur`, or `Indirect`. The only "modifier" so far is the
previous `Rotated`; in the future there will be modifiers such as
"quote" (suppress behaviors so that e.g. the VUI toolbar is not
damaged), "partly moved" (translate the voxels), or "partly destroyed".

The advantages of this redesign are:
* Cloning `Block` values is consistently cheap (which is important for
  transactions).
* It's easy to add more modifiers and have them behave consistently.
* Many modifiers don't create equally deep nesting of pointers and
  thus recursive traversal.

Future work:
* It may make sense to make non-default `BlockAttributes` stored as a
  modifier, instead of part of the `Primitive`.