Test harness for all-is-cubes-gpu (no assertions yet).
I plan to eventually use this for writing individual unit tests for
shader functions. For now, it will at least be a “smoke test” (run the
entire renderer and see if it fails) which we didn't have before.

The test binary will do nothing if GLFW fails to initialize, which is
sufficient to avoid failures on at least GitHub Actions' headless Linux
environment. We might need to revise it in case of an environment which
presents itself as having windowing but not GL.