Add wgpu web glue.
Kludge: wgpu wants a raw-window-handle, and the simplest way to do that
is to invoke winit, but winit wants to take over the canvas size. We
should either use winit fully, or strip it out and do our own
`HasRawWindowHandle` implementation (but if we do the latter then we
will run into conflicts with winit which appears to use a global counter
for "window" ids).