Add a `fmt` subcommand to `xtask`.
This is preparation for using more packages not in the main workspace
(in the same way that `fuzz` is already a separate package). When we do
that, we'll want to have all of the project-wide operations that Cargo
usually provides; hence adding `fmt`.

The utility function `do_for_all_workspaces` will also come in handy.