Fix far/near confusion in camera view frustum.
The previous view_frustum_corners array was used as if far was near and
vice versa. This did little harm since `aab_in_view()` doesn't actually
care (beyond FP error). But, it could have bitten us later.

I replace the array with a struct with named fields, `FrustumPoints`,
so that the type and all its usages specify what the meaning of each
point is.