Precompute view frustum's bounding box.
This gives a -30% improvement in chunk selection costs according to the
benchmark.

Note that in order to preserve the robustness that test
`camera_bad_viewport_doesnt_panic()` was looking for, I had to change
`Viewport::nominal_aspect_ratio` to sanitize its result; otherwise the
matrices would be non-finite and the frustum AAB construction would
panic.