Reduce MAXIMUM_LIGHT_COMPUTATION_COST.
I'm not sure if this is a regression or what, but for the moment, step
time is often exceeding 16 ms and thus greatly reducing frame rate. I'm
currently working on the graphics code and I want to focus on making
that as good as possible, which will be easier if there isn't another
large factor changing the outcomes.

In the long run, we want this to be configurable or auto-adjusted, of
course.