In ChunkedSpaceMesh, replace fixed cost/count limits with time measurement.
We were measuring the time for each block/chunk operation anyway, so
using the times to decide when to stop adds almost no cost, and this
way, we get a more stable frame rate and adapt to the capabilities of
the machine (and future optimization).

This does noticeably reduce chunk updates when e.g. rendering
explosions. But that's also a sign that we need to optimize something
there because those chunk updates are taking up to 20 ms sometimes.