Improve mesh_bench measurement strategy.
Firstly, there is no reason to construct a new `Space` for each
iteration; this makes it less realistic than reuse. Then while fixing
that, I noticed that the benchmark would be including the time to drop
the `Space`, since it was passed by value. So, now all of the benchmarks
use `iter_batched_ref` (even the ones that could use
`iter_with_large_drop`, for consistency) so that the part that is timed
is only the mesh calculation.