Add `Modifier::Move`.
This allows blocks to be displaced out of their cube. While they are
still cropped to the cube grid, a pair of adjacent `Move`s can represent
a block partially in two cubes.

Future work will include automatic animation, so that a `Move` can
all by itself cause a block to slide from one cube to an adjacent one,
and the ability to slide one block out and another in.