Make `UniverseTemplate::build()` produce a Send future.
This requires getting rid of uses of `NoiseFn` modifiers because
they contain `&dyn Noise` which is not `Send`. Instead, I've used some
wrapper functions and some eager conversion to arrays.

In the future, I would like to get rid of `NoiseFn` entirely, but it
seems like there isn't really a good drop-in replacement library that
will serve the same purposes. Perhaps it will be necessary to write our
own noise functions (which has the advantage of aligning with our types
and future needs, and guaranteeing stable implementation as long as we
care).