test-renderers: Add inexact image comparison.
The diff function allows pixels to be displaced from expected positions
(though does not yet have a plan to deal with antialiasing), but the
quantification of difference is still naive; it asks if _any_ pixel
(after position fudging) exceeds a certain value difference. In the
future we may start caring about pixel counts.