wgpu: Implement tone mapping.
The info_text render pipeline is now a "postprocessing" pipeline, runs
unconditionally even when there is no info text, and implements the tone
mapping operator.

The reason to do things that way, rather than the same way I did
previously in the `luminance`-based renderer, is that this should be
more correct in the presence of transparent objects: blending is done
in linear/HDR space, and tone mapping applied only to the final
displayed image. Also, if the TMO is ever expensive, we'll save
computation by deferring it.

And, if I decide to implement bloom, it'll also need at least one
postprocessing pass like this.

A disadvantage of this implementation is that tone mapping doesn't
apply separately to `world` and `ui`; fixing that would require
separate offscreen buffers.