Remove `xaction` dependency.
I was not using any of its functionality, only the reexported `xshell`.
I also added `anyhow::Error` as the error type, instead of
`xaction::Error` which is just a type alias for `Box<dyn Error>`.

Note that this commit does not update to the 0.2 version of `xshell`,
because that has a very different interface that uses a virtual current
directory instead of pushd RAII guards. I might like to switch to that,
but I'm keeping this commit as a simple, low-risk replacement.