Improve `Raycaster::fast_forward`'s choice of t epsilon.
By scaling it according to the ray magnitude, we can make it
block-scaled rather than maybe spanning many blocks, causing the
following stepping to require many seconds of work.

(A tempting thought is “maybe we should normalize the ray”, but then
we'd introduce errors when converting the output back, and it's not
obvious that those errors would be better. Though, probably it wouldn't
matter for realistic cases.)