Update `transparent_one` render test.
The preceding raycaster change apparently *fixed* the raytracer's
handling of the one-cube space, resulting in a color change that
matched the wgpu output. On investigation, I have also decided that the
math ought to be simplified by disabling lighting, so that the color is
simply a transparency blend of the block color and sky color.