GitHub Actions: Test newer Rust and newer deps.
I've occasionally found problems to report with new Rust releases, so
building all-is-cubes on nightly seems like a worthwhile idea as a
continuous canary.

I also added a mode where the lock file is updated, to catch early when
a new allegedly-minor update might have broken something. Someday I
might try to get this to work with `minimal-versions`, too!