Don't await `StagingBelt::recall()`.
I was given this advice by cwfitzgerald -- there is no reason to wait
for the `StagingBelt::recall` unless you want to minimize memory usage,
because it will just allocate more buffers if the old ones aren't ready
for reuse. It *is* necessary to poll the futures it returns, but we now
do that by throwing them in a `FuturesUnordered` and just polling them,
not awaiting them.

This provides a slight observed performance improvement, but mainly, it
allows the rendering functions to stop being `async`, which will make
them easier to integrate into event loops' expectations of drawing
(both desktop and web).

Future work: it would make sense to poll the futures in the background,
but we don't currently have a way for all-is-cubes-gpu to be given an
async executor or for it to be told that it's OK to create threads
(wasm is still single-threaded).

I did try the alternative of removing `StagingBelt` usage entirely, but
that increased CPU frame time from 5-6 ms to >16 ms.