Add `Tool::PushPull`.
This serves to test the `Modifier::MoveBlock` and also is a fun toy.
It does have some noticeable flaws, such as that if you use it twice
you can split up a moving block (though not in a permanent way).
I plan to address that as part of the same problem as limiting space
modification according to game rules (just as some blocks should be
hard or impossible to break, moving blocks might be not movable-again)
or perhaps as multiblock structures (a stacked move in 2 dimensions is
fine as long as the two parts are moved together).

Also, the tool needs an icon.