VUI: Add layout “gravity”. Expand control_bar_bounds.
The “gravity” determines where widgets are positioned when more
space than required is available. (More tweaks are needed; this
is the minimum that will work.)

We now no longer need control_bar_bounds to be only 1 unit tall,
so it can be enlarged to allow for e.g. horizontal and vertical
bars in that corner.