Delete vertex lighting support.
* `GfxVertex::WANTS_LIGHT`
* `GfxVertex::face()`
* `MeshOptions::use_space_light`
* The `lighting` parameter to `GfxVertex::instantiate_vertex()`

This is never used and likely will not be. Vertex lighting cannot
produce smooth lighting (since there isn't enough neighborhood
information to interpolate across a quad) so it isn't very interesting,
even in the context of exporting a mesh where we don't mind baking the
light data into the mesh.