Revise new_hud_space() + HudInputs to support resizing.
In order to be able to recreate the HUD space (or, perhaps, resize it,
if we can make `Space` and `Widget` support that), we need to have the
ingredients to create it available. `HudInputs` is almost that; it just
needed a couple more fields.

This commit is a refactoring with no new functionality.