Move `Grid` from `space` to `math` module.
`Grid`'s uses are now much more than just defining the gridded volume in
which a `Space` has contents (the original idea be), and it is fairly
parallel to `Aab` which is in the `math` module. Also, I plan to rename
`Grid` to `GridAab` so that `Grid` is consistently used as an adjective
meaning “aligned to the voxel grid”, and most of the things named that
way will be in the `math` module.