Force block resolutions to be powers of 2 only, via an enum.
Instead of `type Resolution = u8`, we have `enum Resolution` which
contains the powers of 2 between 1 and 256. This provides some
advantages:

* Modifiers can always subdivide a mesh to their choice of minimum
  desired resolution without issue (e.g. if we had a resolution=9
  that we wanted to apply a `Move` or other process at resolution=16,
  the least common multiple of 9 and 16 is 144, which is bigger than
  we want to implicitly start computing with).
* Zero, which is not a meaningful resolution, can be excluded from the
  valid value set. This eliminates code for handling the zero case.
* Mesh vertex coordinates are nice round floating-point numbers and
  could even be 16-bit fixed-point numbers (given chunk translations).
* 256 is a nice round number to be the maximum resolution. (Previously
  we were stuck with 255, though 256 could have been achieved if we
  went with a `u8` wrapper that effectively added 1 to the value.)

The disadvantage is that now `Resolution` has to be imported and named
everywhere that builds blocks, rather than using integer literals. But
we should not prioritize simple code over world model simplifications.

Also, `resolution as <numeric type>` used to produce the resolution
value and now produces its log2. Unfortunate that we can't simply
prohibit using `as` on an enum (well, except by introducing a dummy
variant with fields).