Replace crosshair with an image.
The immediate motivation was to support a line of development which
would remove support for block resolutions not a power of 2, but it also
reduces the amount of single-purpose code (thus _hopefully_ a marginal
reduction in compile time and future maintenance costs). On the other
hand, we might want to revert this if we find that we want to give the
crosshair variations (e.g. for tool states) that aren't simple
combinations.