wgpu: Enable HDR; when possible, pick a float surface format.
This is sufficient to enable actual HDR output, at least on macOS.
(Testing on other platforms would be good but I don't have them handy.)

There are still issues, though: we don't know what the maximum dynamic
range (or the color space) actually is, so it's not possible to properly
tone-map this result. See <https://github.com/gfx-rs/wgpu/issues/2920>.
However, it's still an improvement on nothing.