Make texture coordinates (points) generic.
They are still `Vector3<f32>` when they they are being calculated for a
particular block, but after `TextureTile::grid_to_texcoord`, they are an
arbitrary type that the vertex type and texture tile type agree on.

This will allow texture allocators to use whatever kind of texture
coordinates they want. In particular, it will be possible to have tiles
correspond to multiple textures (useful for the glTF export and possibly
for allocating more GPU texture space) without needing to assign them
coordinates in 3D Euclidean space. Also, I will be able to try storing
`u16` integer texture coordinates to reduce vertex size.

The most significant disadvantage of this change is that `BlockVertex`
is now a generic type, which complicates casually using it. But, that
is mostly a problem for `all-is-cubes`'s tests and not for other crates,
since the only place `BlockVertex` is usually mentioned is when
implementing `GfxVertex`.