all-is-cubes-server: stop having a copy of the static files.
I've decided to assume that there aren't going to be any embedded-client
builds starting from crates.io, so on that premise, there is no reason
to stash the client/static files within the server directory. Instead,
we can refer to it directly, and when/if it comes time to make a “real”
release version, we can either build with embedding from the workspace,
or bundle the static files separately in whatever installation package.