Add mutation testing in CI.
Not sure how valuable this will be, but I want to give it a try, and
running in CI means I get to see the results without waiting for manual
runs, and isolates any harm from mutated code.

For now, we will be testing only the `all-is-cubes` crate, and excluding
various categories of code such as `fmt()` functions to concentrate on
higher-value signals.