Replace procedural VUI button icons with image files.
The new images do a better job of depicting what they're going to do
than the old two-letter labels did. Additionally, all of the “debug”
buttons have a “D” in their icons.

This does lose one piece of previous functionality: the label color
would change with the button state. I might at some point introduce
color mapping that can tweak that, but for now, this is “physical”
and looks okay.