Fix space_to_image's handling of bounds.
The bounds would be off by 1 on flipped axes, because we were
effectively subtracting and then adding 1, but in different coordinate
systems (subtracting in the image's coordinate system and adding in the
space's). The new implementation fixes this by doing the adding in the
image's coordinates.

This code could still use more test cases.