Add `load_image::ImageAdapter`.
This allows drawing an `image::Image` into a
`embedded_graphics::DrawTarget`. It's not quite a pure adapter for the
two other libraries, though, because it also does the pixel to
`VoxelBrush` conversion that's useful for us and which
`space_to_blocks()` already did.

Also reimplemented `space_to_blocks()` using the adapter.