Add `SpaceBuilder::filled_with()`.
This new method allows specifying the initial contents of a `Space`,
which is occasionally useful. Previously, my plan was to have
`build_*()` methods that took a choice of contents, but I think this
direction is better because it allows more possibilities for partial
builders, and it means that `.build_empty()` can be just `.build()`
(the actual change will be made next).