Replace toolbar selection pointer with an image.
The new image is supposed to still be a pointer but additionally
convey the idea of a mouse, and identify the specific button bound,
thus conveying to new users that they indicate click actions.

There is also a “pressed” state which is not yet used but which I
hope to use to add feedback on how the game is interpreting mouse
inputs.