Make `chunk_distance_squared_for_view` not lose info on adjacent-to-coordinate-plane chunks.
By adding the output struct `Distance` with an extra field, we tie-break
comparisons between distance values that are axis-aligned and ones which
are off by 1, which allows `chunk_distance_squared_for_view` to be used
for chunk sorting as well as distance checks.

The reason to do this is so that a small `ChunkChart`'s chunk list is
a prefix of any larger one, so that chart computation can be cached.