Replace exhibit name text drawing with use of widgets.
This is intended to clear the path for adding subtitles or other details.
It also eliminates the the last use of `draw_text_in_blocks()`.

The disadvantage is that there is more total code now, but it's hopefully
code that more UI and signage will be able to reuse eventually.