Create minimal macOS app bundle.
(This is for local development purposes only; a proper distributable
version would not be referring to files outside the bundle.)

Note that this does not invoke Cargo, but only runs an already-built
binary; this is because the desktop environment very likely does not
have `cargo` in the $PATH, so we'd have to make more unjustified
assumptions to get it to succeed. I might revisit this later.