Add `Fluff` events for sound effects and particles.
Currently, the only use of this at all is that any `ToolError`
resulting from a click will produce an event. In the future, the `Fluff`
mechanism will handle all short-duration sounds and particle effects.

This is intended as the minimal mechanism to get some sound going,
which I can then build on.

Future work:
* Actually producing audio from this data.
* Continuous ambience (which will not be represented by `Fluff`, but
  by another data format produced from stepping).
* Position data for panning / spatial audio.