Give atrium a directional light source — made of blocks.
A better implementation of this would be to have a skybox that
participates in lighting, but that's not yet a thing; the sky is
limited to one color.

Note that the light is very long because this hides the flaws of the
limited light ray propagation angles. Perhaps someday we might think of
a way to implement more efficient light rays and have more of them.