Fix `ChunkedSpaceMesh` counting itself finished even if it timed out.
We now use a simpler strategy: if the chunk update loop exited early
due to timeout, _that_ defines the missingness/incompleteness. If it
finished, then it must be up-to-date. Besides being simpler, this also
fixes a 1-update lag inherent to the previous approach; if the last
update inserted the final needed chunks, then it would still set
`chunks_were_missing` true.