wgpu: Add framework for unit-testing shader functions.
This is distinct from the existing render-test system in that it allows
and requires inserting new shader code, so that individual functions can
be exercised separately from their effect on the rendered output.

It already caught one bug in the tested function: the `modulo()`
function I defined was not correct for `b != 1.0`, which was not
detected since it was only ever used with `b == 1.0`.

Future work:
* Write more tests. The motivation for this framework was to aid
  implementation of GPU raytracing (for complex transparent blocks),
* Make the input data not fixed, if that proves more practical than
  making the functions tested take parameters.
* Optimize the setup costs; consider reusing some of the objects.
  (Waiting on seeing how long the tests take to run with more cases.)