desktop: Add `inner_main()` function for post-window-creation activities.
This does nothing interesting now (except a new logging message), but
eventually it should help with setup activities like being able to
run the universe template _while the window is up, with UI_ rather than
beforehand.

In the longer term we would like to be able to start up the event loop
before there is a session, which will align better with possible
multi-window support and maybe reduce overall coupling, and this sort of
moves away from that because it demands a session object as parameter,
but I think it will help just to start having more code generic over the
loop/window types.