Handle overflow in `GridAab::translate`.
The new behavior is that the `GridAab` is simply truncated to the part
of itself which fits within the representable range. I reviewed the
existing usages and almost all of them are totally fine with this,
because they are all effectively a translation followed by an
intersection with another `GridAab`.

The exception is `GridArray::translate()`. For now, I will have it
panic on overflow.