Remove `Resolution::R256`, making the maximum 128.
The primary motivation for this is to make fewer cases where arbitrary
block definitions will take a long time to evaluate (though we probably
want to add a step limit there too). But also, 1/256 is a pretty tiny
scale visually; I don't expect it to fit in well in otherwise chunky
voxel content. We can always add it back later.