Add save-then-load glTF tests.
I'm not happy with the deps and run time this adds, but we do need more
test coverage.

In the future, these tests should be moved into `all-is-cubes-port` as
unit tests, so that they can be more efficient; that will require moving
the top-level glTF recording logic into the port crate, which should
happen anyway.