Remove `luminance` and `glfw` support.
This is mainly because I want to add substantial new features to the GPU
rendering — such as instanced block rendering and GPU-raytraced
volumetric transparency — without having to implement them twice and be
constrained by `luminance`'s single-threaded and `&mut`-requiring API.

Note that this doesn't strictly imply dropping `glfw`. I've vaguely
heard that it might be possible to use `glfw` windowing with `wgpu`,
but I gave that a try once and couldn't get it to work. If it were
possible, I'd be interested in pursuing it because GLFW seems to have
slightly better behavior than `winit` in some areas of window management
and event handling; specifically I have noticed (on macOS):

* cleaner redraws during resize
* smoother mouselook motion
* correct handling of the app window opening while a fullscreen window
  of another app is in front

However, none of these are critical enough to pursue immediately, and
I can always resurrect the glfw code from version control in the future.