wgpu: Move headless renderer into `all-is-cubes-gpu`.
This means it is public (which makes sense since the `HeadlessRenderer`
trait and `RtRenderer` implementation is public), and that it will be
usable within tests that are in the `all-is-cubes-gpu` crate. In
particular, I want to try writing benchmarks of rendering, and to
write unit tests of shaders (which will need more or less the same kind
of initialization as the headless renderer; and the more we can share,
the less fiddly and fragile special-case setup code we need).