wgpu: Compute perpendicular vectors in vertex shader instead of fragment shader.
I figured this might be a noteworthy performance improvement; it turned
out neutral, but the change seems like like a clarity improvement anyway.

Also, I now know that these vectors have the conventional name “tangent”
and “bitangent”. These are usually used for normal mapping, which is not
what we're doing (yet), but it has a related motivation for existence:
choosing an arbitrary consistent coordinate frame in which to compute
modified light.