Move `Icons` from `vui` to `inv`.
This removes a module interdependency, which might allow for splitting
`vui` into another crate (not yet sure if I want to do that), but in any
case makes sense since `Icons` is closely tied to `Tool`.

In the future we might want a more decoupled approach where `Tool` isn't
aware of any graphical choices at all, but if we do that then that will
likely mean removing `Icons` entirely.