Change `Cursor::face_selected()` to return the block AABB face.
Previously, the selected face was define as the face of the cube
that the ray entered. This definition is unintuitive for flattish
blocks, so it now uses the AABB of the block voxels.

(Note that this is arguably a leak of implementation details, because
currently that AABB is not necessarily minimal. However, since the
cursor behavior only affects player aiming and not any other game
mechanic, I think it's good enough for now.)