Remove `AxisMirrorIter::size_hint()` implementation.
It is never used, because it is only used inside `flat_map()` which
cannot give a size hint. (Perhaps we should add a custom iterator to
fix that, but chunks aren't generally supposed to be collected so it
would not add much benefit.) So, removing this code means faster
compiles and less untested code.