Log overly-long gaps in `YieldProgress` calls.
Note that this is not really an accurate detection of undesired pauses,
In particular, if the code calls other code that does not interact with
`YieldProgress` but *does* yield (e.g. because it is an IO operation),
then the `YieldProgress` will observe a gap in yielding. In order to be
accurate, we would need to use a background task that detects when it is
prevented from running.

But, for now, this will be a useful tool since the current uses of
`YieldProgress` are 100% compute, and it also detects the situation
where the task is proceeding but not _reporting_ that is proceeding
(though that should in principle have a considerably longer timeout than
for yielding).