Add `BlockAttributes::placement_action`.
Part of <https://github.com/kpreid/all-is-cubes/issues/480>; this moves
in the direction of being able to eliminate `Tool` entirely in favor of
`Block`s with extra behaviors.

Missing functionality: The placement action cannot have any custom
interaction with the character's inventory. This should be fixed as part
of generally making `Operation`s more able to interact with inventories.