Create `Primitive::Air`.
This makes the various special cases for `AIR` an explicit part of the
data model, which will avoid the particular definition of `AIR` getting
serialized into save files (once those exist). It's sort of like making
`AIR` a `Primitive::Indirect`, except pointing at a built-in definition
instead of a `URef<Space>`.

Someday we may want to eliminate all of the special cases, but for now,
this is an improvement. Another possibility in the future is to have
_more_ built-in block definitions, such as for UI elements, to reduce
the amount of `Universe`-member creation needed to get the UI running,
or to allow efficient dynamic procedural generation. In that case, we
might generalize `Primitive::Air` to an entire contained enum of
built-in definitions.