Introduce `MinEval`; avoid computing derived properties wastefully.
The innards of `Block::evaluate()` now use `MinEval` as the
representation of blocks through all `Modifier` stages, and don't
convert to `EvaluatedBlock` (and compute derived properties) until the
end. This significantly improves the performance of modifier evaluation,
according to the benchmark I have.

It slightly increases the time to evaluate a modifier-less atom (+25%
time); I have not identified the cause. But since that's microseconds
rather than the milliseconds a voxel block might take, it seems like a
good tradeoff.

The `MinEval` type is private, and there are no changes to the public
API. There may be some edge case behavior changes around resolution-1
voxel blocks.