Add `new()/compute()` split to `BlockMesh` just like `SpaceMesh`.
This isn't used anywhere yet, but it keeps the option open to start
reusing allocations. Unfortunately, `ChunkedSpaceMesh` is currently
benefiting from the ability to compare just-constructed meshes to
previous ones, so we can't just drop this in and get a guaranteed win.
We'll need to test the performance, and perhaps make further changes.