raycast: Add check for extremely large direction vectors.
They could cause fast_forward() to make a poor estimate, thus causing
an unacceptably slow amount of of stepping to find the next cube.
To avoid it, we zero the direction vector so that the raycaster will
only return the origin cube. (Perhaps this should really be an error
instead, but we'd need to make the constructor fallible and all that.)