Stop running CI on *all* pushes; add workflow_dispatch.
This will avoid duplicate runs for every pull request. The original
motivation for `on: push` was so that I could push branches to test them
without creating PRs, but now that I know `workflow_dispatch` exists,
that's a better tool for the job; it being manual isn't a big deal given
how rarely it comes up.