Use `BlockProvider`s to carry button states, with and without label.
This means that we don't need awkward alternate representations like
`[Block; 2]` to carry the concrete button states — it automatically
expands to cover whatever `ToggleButtonVisualState` etc does.

It's not quite as efficient because it's constructing `HashMap`s (and
more of them than strictly necessary), but this should be fine for the
purpose.