Revert 8a8887da Rust dependency updates.
Something changed in `wasm-bindgen` or neighbors broke the web servers,
in that the wasm build no longer produces a `dist/` directory that
contains a wasm module.

Hopefully this is merely an incompatibility between wasm-bindgen
and webpack, because I want to entirely remove npm/webpack from the
build process soon; for now, this will get it working again.